requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
typing_extensions==4.1.1
yarl==1.7.2
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pypng==0.20220715.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
Pillow==9.5.0
httplib2==0.22.0
idna==3.4
cachetools==5.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
typing_extensions==4.6.3
urllib3==1.26.16




